add auto-creating build script
i'm not sure if this will work (if not we'll just add functions manually)

tl;dr it walks the directory on build, scans each file with #[tauri::command] or #[command] and generates a file called `auto_commands.rs`. the main program will import the function `get_cmds()` and tauri will use that to generate handlers.

one downside is you cannot have commands in main.rs. but please, do not have commands in main.rs

please keep auto_commands.rs empty (it is on the gitignore list, no automatically committing it!)

this is kinda hacky, but is the best dynamic solution for a statically, compiled language :)

if it doesn't work long term, i'll rip it out lol